Report missing CA if cert requests are pending
This sets the charm status to blocked to make it obvious when the CA
cert is missing and there are pending cert requests being blocked.

This also moves the optional interfaces checks down to ensure that they
don't mask the more important status messages.  (E.g., if Vault is
providing certs for Etcd, it's more important to know that Vault is
sealed or missing the CA than to know that Etcd doesn't have its cert
yet.)

This also adds some error checking to gracefully handle the case where
Vault becomes sealed after it was successfully started rather than
having it go into a hook error.

Change-Id: I18a5dbeabc562e14d164f82c041fed207032f52b
Closes-Bug: 1840696